Short Description of me
  Final Elevator Pitch:
Hi! I’m David Elijah, a web developer with mastery in the use of Artificial Intelligence Software Tools for Career Development. I acquired an extensive and in-depth training in web development and AI Career Essentials which equipped me with the knowledge to work on several projects like designs for an AI Portfolio, BMI calculator and music for skill exhibition. Though the task of developing a website is stressful, I always believe that failure isn’t an option and I’ll get it done if I put in more effort.